New typeahead controller.
This will validate if the item entered exists on a typeahead list. To be used with a typeahead widget.

Config is driven from fields configuration.

To use add a key of typeahead on the field you want to validate like so:

```
fields: {
  country: {
    typeahead: {
      list: ['England', 'France', 'Poland']
    }
  }
}
```

Where list is a typeahead list you want to validate against. In practice this can be required() in.

The controller will also skip validation if you add a dependency and the criteria isn't met (ala normal fields).