Ensure forking and referring works correctly from the confirm page.
If a fork condition is met we should go down the fork and not back to the confirm page

We do this by checking if a fork has been unlocked from the sessionModel steps.

This fixes a number of bugs where people have gotten to the confirm page. Edited their choices and have managed to go back to the confirm page when they should have been kicked out (due to a fork).